requirements.txt
Flask
Flask-Cors
torch
numpy
ultralytics
Pillow
werkzeug
